tiny-agent-loop/practice/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
openai==2.31.0
packaging==26.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
python-dotenv==1.2.2
requests==2.33.1
serpapi==1.0.2
google_search_results==2.4.2
setuptools==82.0.1
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
